function/requirements.txt
azure-functions
azure-storage-blob==12.18.0
azure-servicebus==7.8.0
pillow
cryptography==43.0.3